enh: improved debugging and error handling, for good and error cases.
Now displaying the process output file if it exists, along with the
standard out, error, and condor log for both good and error cases.
Added in handling of infinite values returned from metrics, and the
case where there are more lines in the scored metric output than there
are labels to be scored.